+++ core/docs/changelog.md
@@ -1,7 +1,7 @@
 _For details about Kurtosis' versioning scheme, as well as how to upgrade, see [the versioning & upgrading page](./versioning-and-upgrading.md)_
 
 # TBD
-* Add .dockerignore file
+* Add .dockerignore file, and a check in `build.sh` to ensure it exists
 
 # 1.4.5
 * Use `alpine` base for the API image & initializer image rather than the `docker` Docker-in-Docker image (which we thought we needed at the start, but don't actually); this saves downloading 530 MB _per CI build_, and so should significantly speed up CI times